Changed the simple mineralsToKeep to ReservedResources structure. Added priority to task (not used now, but will be soon).
Next thing will be to do TaskTrain class and perform the building there.
After this, I can read priority from xml, and prioritise tasks.

In other words it will allow for example when I prioritise Marine range, to not allow to build other things if it could delay the prioritised task.

git-svn-id: https://bwapi.googlecode.com/svn/trunk@373 273ecf2c-f24b-0410-b6a7-e9298bb4873f