Added test cases for attackMove and attackUnit.
Added functions:
Position::isValid - returns true if the position is on the map.
Position::makeValid - modifies the position so that it becomes valid (on the map), moving it at little as possible
TilePosition::makeValid - modifies the tile position so that it becomes valid (on the map), moving it at little as possible

Changed functions:
TilePosition::isValid now returns false for any tile position not on the map, not just for TilePositions::None and TilePositions::Unknown.
Unit command functions now handle positions and tile positions that are off the map by using the nearest valid position or tile position to the given Position.cpp

Bug fixes:
Fixed a bug where Siege Tanks, Goliaths, Reavers, and Carriers would not attack when given an attackUnit command.
Fixed a bug where Starcraft would crash when an AI issued a stop command to a Reaver or Carrier.


git-svn-id: https://bwapi.googlecode.com/svn/trunk@2538 273ecf2c-f24b-0410-b6a7-e9298bb4873f