Remove several blocks of code that were originally required for the multiple onGameEnd calls. I don't know if it behaves much differently than it should, but it appears to me that issues 291 and 299 might have been fixed (tested with dev ai module).
If it's much different from the original then perhaps a revert just to the largest block that was removed in GameInternals (the onMatchEnd event on victory/defeat) with an extra "calledOnMatchEnd" variable that is reset in onGameEnd. Of course, using the value "bool win" so that we don't produce so much repetitive code.

git-svn-id: https://bwapi.googlecode.com/svn/trunk@3154 273ecf2c-f24b-0410-b6a7-e9298bb4873f