Fixed Game::enemy() so that it doesn't return NULL as soon as the enemy leaves the game.
Fixed issue 280. Unit::getUpgradeLevel now works the same way on the server and client - just returns Player::getUpgradeLevel, or 0 if the unit type doesn't use that upgrade type.

Player::getUpgradeLevel returns the upgrade level for the given upgrade if the player is accessible (i.e. self, or complete map info = true), or if the number of completed accessible units owned by that player of a unit type that uses that upgrade type is greater than 0. (makes more sense in code than in english :P).

git-svn-id: https://bwapi.googlecode.com/svn/trunk@3196 273ecf2c-f24b-0410-b6a7-e9298bb4873f